osd/ReplicatedPG: remove broken AccessMode logic
The original intent here was to handle reads in two modes.  For
workloads with read/modify/write ops, the RMW mode would:

 - queue writes for local store and replicas immediately
 - block reads until the write commits to all replicas

For mixed read/write workloads without read/modify/write ops, the
DELAYED mode would:

 - queue writes for replicas
 - allow local reads
 - once replicas commit, queue write locally
 - block local reads until local write completes

In reality, we never use the DELAYED mode.  It's untested and possibly
broken, and it is unlikely we will see a workload where it is important
in the near to mid term.

Signed-off-by: Sage Weil <sage@inktank.com>
Reviewed-by: Samuel Just <sam.just@inktank.com>