osd: add last_user_version to pg_info_t
We add a corresponding user_version to pg_log_entry_t, and the logic
to assign from one to the other and to recover last_user_version from
a master's log. We aren't yet setting it to anything, though.
Notice also that the pg_log_entry_t::user_version is soft state as
the user version isn't required for recovery -- it's simply an assigned
value.

Signed-off-by: Greg Farnum <greg@inktank.com>