ReplicatedPG: add OpContext::user_at_version
Set this up with the existing at_version member, but only increment
it for user_modify ops. Use this when logging the PG's user_version.

Now that we have and are maintaining this PG user version, use it
for the user version on ops that get ENOENT back, and further use
it for the cls_current_version() function. This is a small semantic
change for that function, as previously it would generally return the
same value as the user would get sent back via MOSDOpReply -- but I
don't think it was something you could count on. We now define it as
being the user version of the PG at the start of the op, and as a
bonus it is defined even for read ops (the at_version is only filled
in on write operations).

Signed-off-by: Greg Farnum <greg@inktank.com>