osd: apply changes to user_version/replay_version for backwards compatibility
The system we've been building up works out very nicely for new clients,
but they could not have interoperated with old clients that were only
referring to our replay_version. In order to deal with this, we add
a bad_replay_version to MOSDOpReply which is encoded where we used
to encode replay_version. bad_replay_version follows the same semantics
as reassert_version used to (except that it is filled in on reads), but
is not accessible to new clients, who can see only our properly-controlled
replay_version and user_version. This set of changes means old and
new clients can interoperate correctly when communicating about watches,
etc, because for classic pools the user_version follows the same rules
as the reassert_version did, and we fill in the bad_replay_version with
the user_version.

Signed-off-by: Greg Farnum <greg@inktank.com>