osd: properly set the user_version when using reply_op_error
This function is not just used for error replies; we also use it
when getting replays of completed ops, at which point the user
can legitimately want to see the user_version!
In support of that, we slightly change how ReplicatedPG queries
PGLog to make retrieving both the replay_version and the user_version
easy to get at.

Signed-off-by: Greg Farnum <greg@inktank.com>