osd: fix eval_repop MOSDOpReply generation
We have been setting the replay_version based on the MOSDOp's
reassert_version for a very long time when creating new messages,
which means that the second message (if) sent to clients did not have
the correct version information. Stop filling in replay_version from
the MOSDOp to begin with, and specify the full set of versions
when creating the reply.

Signed-off-by: Greg Farnum <greg@inktank.com>