mon: OSDMonitor: don't propose on prepare_pool_op()
Except in very special cases, we should let PaxosService take its course
and trigger the proposals itself.  In this case, we were proposing right
before returning to PaxosService, and we were returning false on top of it
(most likely to guarantee that PaxosService wouldn't try to propose).

This doesn't make much sense, so let's do it like all the other cool kids
are doing and let PaxosService decide what's best for us.

Signed-off-by: Joao Eduardo Luis <jecluis@gmail.com>