mon: OSDMonitor: split 'osd pool set' out of 'prepare_command'
We should start doing this across the whole 'prepare_command' function.
Makes it prettier to the reader, and easier to add new code.

Change the command to send a string instead of an int to allow us to have
non-integer pool paramters that can be modified.  Support input json with
both int and string values so that we work with all flavors of client.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>