test: Use a portable syntax for seq(1)
Use a portable syntax for seq(1).  GNU seq has a default INCR of 1, but
BSD seq has a default INCR of either +1 or -1, depending on the other
arguments.  INCR must be explicitly specified for portability.

This bug is the reason that I was running into the segfaults whose
fix I reported as BUG #6510.

Signed-off-by: Alan Somers <asomers@gmail.com>
Signed-off-by: Noah Watkins <noahwatkins@gmail.com>