Put machine_type in config.yaml
Reason being, if --worker is passed to teuthology-schedule, we are
already placing the job in the tube for that machine type. If the job
config doesn't specify a machine_type, it will later be assigned to
'plana' by teuthology-run without actually modifying the config. This
can result in jobs running on hardware other than what was requested.

Signed-off-by: Zack Cerza <zack.cerza@inktank.com>