mon/PGMonitor: set floor below which we do not warn about objects/pg
If a cluster has very few objects, do not generate warnings when the
objects/pg for a pool diverges from the cluster average.  This avoids
spurious errors when you have a relatively empty cluster and a lone pool
with a modest number of objects is too far off the (mostly meaningless)
cluster-wide average.

Also include a per-pool min so we ignore mostly-empty pools.

Fixes: #6521
Signed-off-by: Sage Weil <sage@inktank.com>