ceph_manager: workaround for 6116
This is an annoying race, we really should delay going
clean until the backfill peer has acknoledged the clean
info, but we currently don't.  In order to prevent this
bug from messing up the nightlies, we'll delay killing
the peer for 20s to make it likely that the backfill
peer has gotten the clean info.

Workaround: #6116
Signed-off-by: Samuel Just <sam.just@inktank.com>